requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk
google.generativeai
grpcio
pillow
requests
torch==2.3.1
faster-whisper
opencc
vertexai
pydub
git+https://github.com/ozgur/python-firebase
